Module Rename
- Rename PMC module to be more generic instead of importing specific
- Update Read Me for the module and the main project since the name was also changed
