+++ README.md
@@ -1,3 +1,10 @@
-# Page tab processor
+# BioStudies Backend Services
 
-Main project which helps to process and generate page tab xml/json/csv formatted files and read from database.
+Project containing the BioStudies backend services:
+
+* **Submitter**:
+Module that helps to process and generate page tab xml/json/csv formatted files and read from database.
+
+* **Scheduler**:
+Module that contains scheduled task to perform BioStudies related processes like submissions export to the UI or
+Europe-PMC submissions processing.
