--- scheduler/pmc-importer-task/README.md
@@ -1,2 +1,0 @@
-# Pmc-importer-task
-Process european PMC files and import into mongo database.
