--- scheduler/pmc-importer-task/src/main/kotlin/ac/uk/ebi/pmc/importer/data/MongoRepository.kt
@@ -1,40 +1,0 @@
-package ac.uk.ebi.pmc.importer.data
-
-import ac.uk.ebi.pmc.importer.data.docs.ErrorDoc
-import ac.uk.ebi.pmc.importer.data.docs.FileDoc
-import ac.uk.ebi.pmc.importer.data.docs.SubmissionDoc
-import com.mongodb.async.client.MongoClient
-import com.mongodb.client.model.Filters
-import org.bson.types.ObjectId
-import org.litote.kmongo.coroutine.findOne
-import org.litote.kmongo.coroutine.insertOne
-import java.io.File
-
-private const val SUBMISSION_COLLECTION = "submissions"
-private const val ERRORS_COLLECTION = "errors"
-private const val FILES_COLLECTION = "files"
-
-class MongoRepository(
-    private val dataBase: String,
-    private val mongoClient: MongoClient
-) {
-
-    suspend fun save(submissionDoc: SubmissionDoc) {
-        val database = mongoClient.getDatabase(dataBase)
-        val collection = database.getCollection(SUBMISSION_COLLECTION, SubmissionDoc::class.java)
-        collection.insertOne(submissionDoc)
-    }
-
-    suspend fun save(errorDoc: ErrorDoc) {
-        val database = mongoClient.getDatabase(dataBase)
-        val collection = database.getCollection(ERRORS_COLLECTION, ErrorDoc::class.java)
-        collection.insertOne(errorDoc)
-    }
-
-    suspend fun saveSubFile(file: File, accNo: String): ObjectId {
-        val database = mongoClient.getDatabase(dataBase)
-        val collection = database.getCollection(FILES_COLLECTION, FileDoc::class.java)
-        collection.insertOne(FileDoc(file.name, file.absolutePath, accNo))
-        return collection.findOne(Filters.eq("path", file.absolutePath))!!.id
-    }
-}
