--- scheduler/pmc-importer-task/src/main/kotlin/ac/uk/ebi/pmc/importer/data/docs/ErrorDoc.kt
@@ -1,11 +1,0 @@
-package ac.uk.ebi.pmc.importer.data.docs
-
-import java.time.Instant
-
-data class ErrorDoc(
-    val id: String,
-    val body: String,
-    val sourceFile: String,
-    val error: String,
-    val uploaded: Instant = Instant.now()
-)
