--- scheduler/pmc-importer-task/src/main/kotlin/ac/uk/ebi/pmc/importer/data/docs/FilesProcessedDoc.kt
@@ -1,8 +1,0 @@
-package ac.uk.ebi.pmc.importer.data.docs
-
-import java.time.Instant
-
-data class FilesProcessedDoc(
-    val name: String,
-    val uploaded: Instant = Instant.now()
-)
