--- scheduler/pmc-importer-task/src/main/kotlin/ac/uk/ebi/pmc/importer/data/docs/SubmissionDoc.kt
@@ -1,13 +1,0 @@
-package ac.uk.ebi.pmc.importer.data.docs
-
-import org.bson.types.ObjectId
-import java.time.Instant
-
-data class SubmissionDoc(
-    val id: String,
-    val body: String,
-    val sourceFile: String,
-    val files: List<ObjectId>,
-    val uploaded: Instant = Instant.now(),
-    val imported: Boolean = false
-)
