--- scheduler/pmc-importer-task/src/main/kotlin/ac/uk/ebi/pmc/importer/data/inputs/FileInfo.kt
@@ -1,3 +1,0 @@
-package ac.uk.ebi.pmc.importer.data.inputs
-
-data class FileInfo(val filename: String, val accNo: String, val md5: String)
