--- scheduler/pmc-importer-task/src/main/kotlin/ac/uk/ebi/pmc/importer/utils/ProcessUtils.kt
@@ -1,27 +1,0 @@
-package ac.uk.ebi.pmc.importer.utils
-
-import kotlinx.coroutines.delay
-import mu.KotlinLogging
-
-private val logger = KotlinLogging.logger {}
-
-suspend fun <T> retry(
-    times: Int = 3,
-    initialDelay: Long = 100,
-    maxDelay: Long = 1000,
-    factor: Double = 2.0,
-    block: suspend () -> T
-): T {
-    var currentDelay = initialDelay
-    repeat(times - 1) {
-        try {
-            return block()
-        } catch (e: Exception) {
-            logger.error { e }
-        }
-        delay(currentDelay)
-        currentDelay = (currentDelay * factor).toLong().coerceAtMost(maxDelay)
-    }
-
-    return block()
-}
