+++ scheduler/pmc-processor-task/README.md
@@ -1,0 +1,3 @@
+# Pmc-processor-task
+- Process european PMC files and import into mongo database.
+- Uses processed files to make the corresponding submissions to BioStudies database
