+++ scheduler/pmc-processor-task/src/main/kotlin/ac/uk/ebi/pmc/data/docs/ErrorDoc.kt
@@ -1,0 +1,11 @@
+package ac.uk.ebi.pmc.data.docs
+
+import java.time.Instant
+
+data class ErrorDoc(
+    val id: String,
+    val body: String,
+    val sourceFile: String,
+    val error: String,
+    val uploaded: Instant = Instant.now()
+)
