+++ scheduler/pmc-processor-task/src/main/kotlin/ac/uk/ebi/pmc/data/docs/FilesProcessedDoc.kt
@@ -1,0 +1,8 @@
+package ac.uk.ebi.pmc.data.docs
+
+import java.time.Instant
+
+data class FilesProcessedDoc(
+    val name: String,
+    val uploaded: Instant = Instant.now()
+)
