+++ scheduler/pmc-processor-task/src/main/kotlin/ac/uk/ebi/pmc/data/inputs/FileInfo.kt
@@ -1,0 +1,3 @@
+package ac.uk.ebi.pmc.data.inputs
+
+data class FileInfo(val filename: String, val accNo: String, val md5: String)
