+++ scheduler/pmc-processor-task/src/main/kotlin/ac/uk/ebi/pmc/utils/ProcessUtils.kt
@@ -1,0 +1,27 @@
+package ac.uk.ebi.pmc.utils
+
+import kotlinx.coroutines.delay
+import mu.KotlinLogging
+
+private val logger = KotlinLogging.logger {}
+
+suspend fun <T> retry(
+    times: Int = 3,
+    initialDelay: Long = 100,
+    maxDelay: Long = 1000,
+    factor: Double = 2.0,
+    block: suspend () -> T
+): T {
+    var currentDelay = initialDelay
+    repeat(times - 1) {
+        try {
+            return block()
+        } catch (e: Exception) {
+            logger.error { e }
+        }
+        delay(currentDelay)
+        currentDelay = (currentDelay * factor).toLong().coerceAtMost(maxDelay)
+    }
+
+    return block()
+}
